PcgenFtlTestCase: convert to JUnit 4 (#4747)
