+++ code/src/test/pcgen/inttest/game_35e/pcGenGUIAliceTest.java
@@ -1,47 +1,17 @@
 
 package pcgen.inttest.game_35e;
 
-import junit.framework.Test;
-import junit.framework.TestSuite;
 import pcgen.inttest.PcgenFtlTestCase;
+
+import org.junit.Test;
 
 /**
  * Tests a Unit Test Case designed to hit many features of PCGen
  * See the PCG file for details
  */
-@SuppressWarnings("nls")
 public class pcGenGUIAliceTest extends PcgenFtlTestCase
 {
-
-
-	public pcGenGUIAliceTest()
-	{
-		super("35e_alice");
-	}
-
-	/**
-	 * standard JUnit style constructor
-	 * 
-	 * @param name
-	 */
-	public pcGenGUIAliceTest(String name)
-	{
-		super(name);
-	}
-
-	/**
-	 * @return A <tt>TestSuite</tt>
-	 */
-	public static Test suite()
-	{
-		return new TestSuite(pcGenGUIAliceTest.class);
-	}
-
-	/**
-	 * Loads and outputs the character.
-	 * 
-	 * @throws Exception If an error occurs.
-	 */
+	@Test
 	public void testCode() throws Exception
 	{
 		runTest("35e_Alice", "35e");
