+++ code/src/test/pcgen/inttest/game_35e/pcGenGUICloudGiantTest.java
@@ -1,50 +1,19 @@
 
 package pcgen.inttest.game_35e;
 
-import pcgen.inttest.PcgenFtlTestCase;
-
 import java.io.IOException;
 
-import junit.framework.Test;
-import junit.framework.TestSuite;
+import pcgen.inttest.PcgenFtlTestCase;
+
+import org.junit.Test;
 
 /**
  * Tests a 3.5e Cloud Giant with a Half-Dragon (Brass) template applied.
  * See it's PCG file for what it contains.  
  */
-@SuppressWarnings("nls")
 public class pcGenGUICloudGiantTest extends PcgenFtlTestCase
 {
-
-
-	public pcGenGUICloudGiantTest()
-	{
-		super("35e_cloudgiant");
-	}
-
-	/**
-	 * standard JUnit style constructor
-	 * @param name
-	 */
-	public pcGenGUICloudGiantTest(String name)
-	{
-		super(name);
-	}
-
-	/**
-	 * return the test suite for this test
-	 * @return the test suite for this test
-	 */
-	public static Test suite()
-	{
-		return new TestSuite(pcGenGUICloudGiantTest.class);
-	}
-
-	/**
-	 * Run the test.
-	 *
-	 * @throws IOException Signals that an I/O exception has occurred.
-	 */
+	@Test
 	public void testCloudGiantHalfDragon() throws IOException
 	{
 		runTest("35e_CloudGiantHalfDragon", "35e");
