+++ code/src/test/pcgen/inttest/game_35e/pcGenGUIDaveTest.java
@@ -1,47 +1,16 @@
 
 package pcgen.inttest.game_35e;
-
 import pcgen.inttest.PcgenFtlTestCase;
-import junit.framework.Test;
-import junit.framework.TestSuite;
+
+import org.junit.Test;
 
 /**
  * Tests a Unit Test Case designed to hit many features of PCGen
  * See the PCG file for details
  */
-@SuppressWarnings("nls")
 public class pcGenGUIDaveTest extends PcgenFtlTestCase
 {
-
-
-	public pcGenGUIDaveTest()
-	{
-		super("35e_dave");
-	}
-
-	/**
-	 * standard JUnit style constructor
-	 * 
-	 * @param name
-	 */
-	public pcGenGUIDaveTest(String name)
-	{
-		super(name);
-	}
-
-	/**
-	 * @return A <tt>TestSuite</tt>
-	 */
-	public static Test suite()
-	{
-		return new TestSuite(pcGenGUIDaveTest.class);
-	}
-
-	/**
-	 * Loads and outputs the character.
-	 * 
-	 * @throws Exception If an error occurs.
-	 */
+	@Test
 	public void testCode() throws Exception
 	{
 		runTest("35e_Dave", "35e");
