+++ code/src/test/pcgen/inttest/game_35e/pcGenGUIFranTest.java
@@ -2,46 +2,16 @@
 package pcgen.inttest.game_35e;
 
 import pcgen.inttest.PcgenFtlTestCase;
-import junit.framework.Test;
-import junit.framework.TestSuite;
+
+import org.junit.Test;
 
 /**
  * Tests a Unit Test Case designed to hit many features of PCGen
  * See the PCG file for details
  */
-@SuppressWarnings("nls")
 public class pcGenGUIFranTest extends PcgenFtlTestCase
 {
-
-
-	public pcGenGUIFranTest()
-	{
-		super("35e_fran");
-	}
-
-	/**
-	 * standard JUnit style constructor
-	 * 
-	 * @param name
-	 */
-	public pcGenGUIFranTest(String name)
-	{
-		super(name);
-	}
-
-	/**
-	 * @return A <tt>TestSuite</tt>
-	 */
-	public static Test suite()
-	{
-		return new TestSuite(pcGenGUIFranTest.class);
-	}
-
-	/**
-	 * Loads and outputs the character.
-	 * 
-	 * @throws Exception If an error occurs.
-	 */
+	@Test
 	public void testCode() throws Exception
 	{
 		runTest("35e_Fran", "35e");
