+++ code/src/test/pcgen/inttest/game_35e/pcGenGUIIvanTest.java
@@ -2,46 +2,16 @@
 package pcgen.inttest.game_35e;
 
 import pcgen.inttest.PcgenFtlTestCase;
-import junit.framework.Test;
-import junit.framework.TestSuite;
+
+import org.junit.Test;
 
 /**
  * Tests a Unit Test Case designed to hit many features of PCGen
  * See the PCG file for details
  */
-@SuppressWarnings("nls")
 public class pcGenGUIIvanTest extends PcgenFtlTestCase
 {
-
-
-	public pcGenGUIIvanTest()
-	{
-		super("35e_ivan");
-	}
-
-	/**
-	 * standard JUnit style constructor
-	 * 
-	 * @param name
-	 */
-	public pcGenGUIIvanTest(String name)
-	{
-		super(name);
-	}
-
-	/**
-	 * @return A <tt>TestSuite</tt>
-	 */
-	public static Test suite()
-	{
-		return new TestSuite(pcGenGUIIvanTest.class);
-	}
-
-	/**
-	 * Loads and outputs the character.
-	 * 
-	 * @throws Exception If an error occurs.
-	 */
+	@Test
 	public void testCode() throws Exception
 	{
 		runTest("35e_Ivan", "35e");
