+++ code/src/test/pcgen/inttest/game_modern/pcGenGUIIlyanaTest.java
@@ -20,46 +20,16 @@
 package pcgen.inttest.game_modern;
 
 import pcgen.inttest.PcgenFtlTestCase;
-import junit.framework.Test;
-import junit.framework.TestSuite;
+
+import org.junit.Test;
 
 /**
  * Tests a Modern Tough Hero 3/Strong Hero 3.
  * See the PCG file for details
  */
-@SuppressWarnings("nls")
 public class pcGenGUIIlyanaTest extends PcgenFtlTestCase
 {
-
-
-	public pcGenGUIIlyanaTest()
-	{
-		super("msrd_ilyana");
-	}
-
-	/**
-	 * standard JUnit style constructor
-	 * 
-	 * @param name
-	 */
-	public pcGenGUIIlyanaTest(String name)
-	{
-		super(name);
-	}
-
-	/**
-	 * @return A <tt>TestSuite</tt>
-	 */
-	public static Test suite()
-	{
-		return new TestSuite(pcGenGUIIlyanaTest.class);
-	}
-
-	/**
-	 * Loads and outputs the character.
-	 * 
-	 * @throws Exception If an error occurs.
-	 */
+	@Test
 	public void testIlyana() throws Exception
 	{
 		runTest("msrd_Ilyana", "Modern");
